added Zenodo AI fecth, Drive mask fetch
Zenodo is now utilized for storing the .pth files. The utilities/zenodo_link function reads the zenodo_UNETR.txt file to extract the DOI and checks if the file exists in a local folder. If it doesn't, the pipelines\fetch_dl_models function downloads the file from Zenodo.

Before employing AI to generate kidney masks, the pipeline/fetch_kidney_masks function attempts to locate the manually generated masks in Google Drive. If none are found, AI is employed to generate the masks using the pipelines\segment_kidneys function.